requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=dev.txt dev.in
#
appdirs==1.4.4            # via virtualenv
attrs==19.3.0             # via -r test.txt, jsonschema, pytest
backcall==0.1.0           # via ipython
bleach==3.1.5             # via nbconvert
cfgv==3.1.0               # via pre-commit
click==7.1.2              # via pip-tools
coverage==5.1             # via -r test.txt, pytest-cov
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
distlib==0.3.0            # via virtualenv
entrypoints==0.3          # via -r test.txt, flake8, nbconvert
filelock==3.0.12          # via tox, virtualenv
flake8==3.7.9             # via -r test.txt
identify==1.4.15          # via pre-commit
ipykernel==5.2.1          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.14.0           # via -r dev.in, ipykernel, ipywidgets, jupyter-console
ipywidgets==7.5.1         # via jupyter
jedi==0.17.0              # via ipython
jinja2==2.11.2            # via nbconvert, notebook
jsonschema==3.2.0         # via nbformat
jupyter-client==6.1.3     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==6.1.0    # via jupyter
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0            # via -r dev.in
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via -r test.txt, flake8
mistune==0.8.4            # via nbconvert
more-itertools==8.2.0     # via -r test.txt, pytest
nbconvert==5.6.1          # via jupyter, notebook
nbformat==5.0.6           # via ipywidgets, nbconvert, notebook
nodeenv==1.3.5            # via pre-commit
notebook==6.0.3           # via jupyter, widgetsnbextension
packaging==20.3           # via -r test.txt, bleach, pytest, tox
pandocfilters==1.4.2      # via nbconvert
parso==0.7.0              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.1.2          # via -r dev.in
pluggy==0.13.1            # via -r test.txt, pytest, tox
pre-commit==2.4.0         # via -r dev.in
prometheus-client==0.7.1  # via notebook
prompt-toolkit==3.0.5     # via ipython, jupyter-console
ptyprocess==0.6.0         # via pexpect, terminado
py==1.8.1                 # via -r test.txt, pytest, tox
pycodestyle==2.5.0        # via -r test.txt, flake8
pydevd_pycharm==201.7223.92  # via -r dev.in, -r test.txt
pyflakes==2.1.1           # via -r test.txt, flake8
pygments==2.6.1           # via ipython, jupyter-console, nbconvert, qtconsole
pyparsing==2.4.7          # via -r test.txt, packaging
pyrsistent==0.16.0        # via jsonschema
pytest-cov==2.8.1         # via -r test.txt
pytest-mock==3.1.0        # via -r test.txt
pytest-reorder==0.1.1     # via -r test.txt
pytest==5.4.2             # via -r test.txt, pytest-cov, pytest-mock, pytest-reorder
python-dateutil==2.8.1    # via jupyter-client
pyyaml==5.3.1             # via pre-commit
pyzmq==19.0.1             # via jupyter-client, notebook, qtconsole
qtconsole==4.7.3          # via jupyter
qtpy==1.9.0               # via qtconsole
send2trash==1.5.0         # via notebook
six==1.14.0               # via -r test.txt, bleach, jsonschema, packaging, pip-tools, python-dateutil, tox, tox-pip-extensions, traitlets, virtualenv
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
toml==0.10.0              # via pre-commit, tox
tornado==6.0.4            # via ipykernel, jupyter-client, notebook, terminado
tox-pip-extensions==1.6.0  # via -r dev.in
tox==3.15.0               # via -r dev.in, tox-pip-extensions
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
virtualenv==20.0.20       # via pre-commit, tox
wcwidth==0.1.9            # via -r test.txt, prompt-toolkit, pytest
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
